Move TritonOperator to operator module
